requirements.txt
streamlit
langchain
huggingface_hub
faiss-cpu
tiktoken
unstructured
pypdf
python-dotenv
